commit to pull

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -2,9 +2,11 @@ package com.techelevator.tenmo.dao;
 
 import com.techelevator.tenmo.model.Account;
 
+import java.math.BigDecimal;
+
 public interface AccountDao {
 
-    int findBalanceById(int userId);
+    BigDecimal findBalanceById(int userId);
     boolean addBalance(int balanceToAdd, int userId);
     boolean subtractBalance(int balanceToSubtract, int userId);
     Account findAccountById(int userId);
